NFC tracker: promote to a full module directory
And break out the invariant checking and snapshot generation logic to
their own files.

Closes #5.

TESTED=it builds.
